requirements.txt
altair
numpy
pandas
pydeck
streamlit
PyPDF2
pinecone-client
langchain
openai
sendgrid
markdown
markdown
mistune
replicate
langchain_community
langchain_core
langchain-openai
pydantic